LayoutTests/imported/w3c/web-platform-tests/xhr/json.any.worker-expected.txt


PASS Ensure the correct JSON parser is used
FAIL Ensure UTF-16 results in an error assert_equals: expected null but got object "[object Object]"
